+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 Log of changes since the 2.0 version
 
 
+### 0.64.1
+- Update changeset-map to 1.4.2
+- Add Bing as background layer option
+- Add copy to clipboard button on filter header
+
 ### 0.64.0
 - Update URLs to osmcha.org